+++ test/util/model_utils/from_json_date_millis_test.dart
@@ -1,5 +1,4 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:folly_fields/util/duplet.dart';
 import 'package:folly_fields/util/model_utils.dart';
 
 ///
@@ -12,28 +11,24 @@ void main() {
     () {
       final DateTime now = DateTime.now().copyWith(microsecond: 0);
 
-      final Map<Duplet<int?, DateTime?>, dynamic> domain =
-          <Duplet<int?, DateTime?>, dynamic>{
-        Duplet<int?, DateTime?>(null, now): now,
-        Duplet<int?, DateTime?>(now.millisecondsSinceEpoch, null): now,
-        Duplet<int?, DateTime?>(8640000000000001, now): now,
-        const Duplet<int?, DateTime?>(null, null): isA<DateTime>(),
-        const Duplet<int?, DateTime?>(-1, null): isA<DateTime>(),
-        const Duplet<int?, DateTime?>(0, null): DateTime.utc(1970).toLocal(),
-        const Duplet<int?, DateTime?>(1, null):
-            DateTime.utc(1970, 1, 1, 0, 0, 0, 1).toLocal(),
-        const Duplet<int?, DateTime?>(1999, null):
-            DateTime.utc(1970, 1, 1, 0, 0, 1, 999).toLocal(),
-        const Duplet<int?, DateTime?>(59999, null):
-            DateTime.utc(1970, 1, 1, 0, 0, 59, 999).toLocal(),
-        const Duplet<int?, DateTime?>(3599999, null):
+      final Map<({int? a, DateTime? b}), dynamic> domain =
+          <({int? a, DateTime? b}), dynamic>{
+        (a: null, b: now): now,
+        (a: now.millisecondsSinceEpoch, b: null): now,
+        (a: 8640000000000001, b: now): now,
+        (a: null, b: null): isA<DateTime>(),
+        (a: -1, b: null): isA<DateTime>(),
+        (a: 0, b: null): DateTime.utc(1970).toLocal(),
+        (a: 1, b: null): DateTime.utc(1970, 1, 1, 0, 0, 0, 1).toLocal(),
+        (a: 1999, b: null): DateTime.utc(1970, 1, 1, 0, 0, 1, 999).toLocal(),
+        (a: 59999, b: null): DateTime.utc(1970, 1, 1, 0, 0, 59, 999).toLocal(),
+        (a: 3599999, b: null):
             DateTime.utc(1970, 1, 1, 0, 59, 59, 999).toLocal(),
-        const Duplet<int?, DateTime?>(8640000000000000, null):
-            DateTime.utc(275760, 9, 13).toLocal(),
-        const Duplet<int?, DateTime?>(8640000000000001, null): isA<DateTime>(),
+        (a: 8640000000000000, b: null): DateTime.utc(275760, 9, 13).toLocal(),
+        (a: 8640000000000001, b: null): isA<DateTime>(),
       };
 
-      for (final MapEntry<Duplet<int?, DateTime?>, dynamic> input
+      for (final MapEntry<({int? a, DateTime? b}), dynamic> input
           in domain.entries) {
         test(
           '${input.key} // ${input.value}',