+++ test/util/model_utils/from_json_decimal_double_test.dart
@@ -1,6 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:folly_fields/util/decimal.dart';
-import 'package:folly_fields/util/duplet.dart';
 import 'package:folly_fields/util/model_utils.dart';
 
 ///
@@ -11,24 +10,20 @@ void main() {
   group(
     'ModelUtils fromJsonDecimalDouble',
     () {
-      final Map<Duplet<double?, int>, Decimal> domain =
-          <Duplet<double?, int>, Decimal>{
-        const Duplet<double?, int>(null, 0): Decimal(precision: 0, intValue: 0),
-        const Duplet<double?, int>(1, 0): Decimal(precision: 0, intValue: 1),
-        const Duplet<double?, int>(1.1, 0): Decimal(precision: 0, intValue: 1),
-        const Duplet<double?, int>(1.12, 0): Decimal(precision: 0, intValue: 1),
-        const Duplet<double?, int>(1.123, 0):
-            Decimal(precision: 0, intValue: 1),
-        const Duplet<double?, int>(null, 2): Decimal(precision: 2, intValue: 0),
-        const Duplet<double?, int>(1.1, 2):
-            Decimal(precision: 2, intValue: 110),
-        const Duplet<double?, int>(1.12, 2):
-            Decimal(precision: 2, intValue: 112),
-        const Duplet<double?, int>(1.123, 2):
-            Decimal(precision: 2, intValue: 112),
+      final Map<({double? a, int b}), Decimal> domain =
+          <({double? a, int b}), Decimal>{
+        (a: null, b: 0): Decimal(precision: 0, intValue: 0),
+        (a: 1, b: 0): Decimal(precision: 0, intValue: 1),
+        (a: 1.1, b: 0): Decimal(precision: 0, intValue: 1),
+        (a: 1.12, b: 0): Decimal(precision: 0, intValue: 1),
+        (a: 1.123, b: 0): Decimal(precision: 0, intValue: 1),
+        (a: null, b: 2): Decimal(precision: 2, intValue: 0),
+        (a: 1.1, b: 2): Decimal(precision: 2, intValue: 110),
+        (a: 1.12, b: 2): Decimal(precision: 2, intValue: 112),
+        (a: 1.123, b: 2): Decimal(precision: 2, intValue: 112),
       };
 
-      for (final MapEntry<Duplet<double?, int>, Decimal> input
+      for (final MapEntry<({double? a, int b}), Decimal> input
           in domain.entries) {
         test(
           '${input.key} // ${input.value}',